Got it compiling and the os check and dll loading for the windows api working! I have to fix some features of the @os block that are not yet supported but will be soon hopefully. Current error we are getting when trygin to compile it with the io lib
error[@os Error]: @os block requires a target OS to be set (--target-os
flag)
  --> C:\Users\conno\Desktop\projects_and_software\Lux\std\io.lx:108:6
     |
 108 |
     |      ^
     |

error[Function Body Error]: Function 'local_write' body failed
typechecking
  --> C:\Users\conno\Desktop\projects_and_software\Lux\std\io.lx:107:22
     |
 107 |
     |                      ^
     |
I believe it is spanning from an @os block being inside of a function
and the tc not understanding how it is supose to handle and or work with
it, and I am too tired to fix this. It is currently 4:15am EST time.

+++ src/lsp/lsp_symbols.c
@@ -31,7 +31,7 @@ static int find_token_for_name(LSPDocument *doc, const char *name) {
   size_t name_len = strlen(name);
   for (size_t i = 0; i < doc->token_count; i++) {
     Token *tok = &doc->tokens[i];
-    if (tok->type_ == TOK_IDENTIFIER && tok->length == name_len &&
+    if (tok->type_ == TOK_IDENTIFIER && tok->length == (size_t)name_len &&
         tok->value && strncmp(tok->value, name, name_len) == 0) {
       return (int)i;
     }
